+++ R/globals.R
@@ -1,2 +1,2 @@
 
-utils::globalVariables(c(".", "an.error.occured"))
+utils::globalVariables(c(".", "an.error.occured", "pb"))