+++ R/worldfootballr_helpers.R
@@ -111,4 +111,99 @@ fb_player_urls <- function(team_url) {
 }
 
 
+#' Get transfermarkt Team URLs
+#'
+#' Returns the URLs for all teams for a given league season
+#'
+#' @param country_name the country of the league's players
+#' @param start_year the start year of the season (2020 for the 20/21 season)
+#' @param league_url league url from transfermarkt.com. To be used when country_name not avalilable in main function
+#'
+#' @return returns a character vector of all transfermarkt team URLs for a selected league
+#'
+#' @importFrom magrittr %>%
+#' @importFrom rlang .data
+#' @importFrom utils read.csv
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' team_urls <- tm_league_team_urls(country_name = "England", start_year = 2020)
+#' }
+tm_league_team_urls <- function(country_name, start_year, league_url = NA) {
+  main_url <- "https://www.transfermarkt.com"
+
+  if(is.na(league_url)) {
+    meta_df <- read.csv(url("https://raw.githubusercontent.com/JaseZiv/worldfootballR_data/master/raw-data/transfermarkt_leagues/main_comp_seasons.csv"),
+                        stringsAsFactors = F)
+
+    tryCatch({meta_df_seasons <- meta_df %>%
+      dplyr::filter(.data$country %in% country_name, .data$season_start_year %in% start_year)}, error = function(e) {meta_df_seasons <- data.frame()})
+
+    if(nrow(meta_df_seasons) == 0) {
+      stop(glue::glue("Country {country_name} or season {start_year} not found. Check that the country and season exists at https://github.com/JaseZiv/worldfootballR_data/blob/master/raw-data/transfermarkt_leagues/main_comp_seasons.csv"))
+    }
+
+    season_url <- meta_df_seasons$season_urls
+
+  } else {
+    tryCatch({league_page <- xml2::read_html(league_url)}, error = function(e) {league_page <- c()})
+
+    tryCatch({country_name <- league_page %>%
+      rvest::html_nodes(".profilheader") %>%
+      rvest::html_node("img") %>%
+      rvest::html_attr("alt") %>% .[!is.na(.)]}, error = function(e) {country_name <- NA_character_})
+
+    if(length(league_page) == 0) {
+      stop(glue::glue("League URL(s) {league_url} not found. Please check transfermarkt.com for the correct league URL"))
+    }
+
+    season_url <- paste0(league_url, "?saison_id=", start_year)
+
+  }
+
+  season_page <- xml2::read_html(season_url)
+
+  team_urls <- season_page %>%
+    rvest::html_nodes(".items") %>%
+    rvest::html_nodes(".vereinprofil_tooltip") %>% rvest::html_attr("href") %>%
+    unique() %>% paste0(main_url, .)
+
+  return(team_urls)
+}
+
+
+#' Get transfermarkt Player URLs
+#'
+#' Returns the transfermarkt URLs for all players for a given team
+#'
+#' @param team_url the player's team URL (can be from tm_league_team_urls())
+#'
+#' @return returns a character vector of all transfermarkt player URLs for a selected team
+#'
+#' @importFrom magrittr %>%
+#' @importFrom rlang .data
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' team_url <- "https://www.transfermarkt.com/fc-burnley/startseite/verein/1132/saison_id/2020"
+#' tm_team_player_urls(team_url = team_url)
+#' }
+tm_team_player_urls <- function(team_url) {
+
+  main_url <- "https://www.transfermarkt.com"
+
+  tryCatch({team_page <- xml2::read_html(team_url)}, error = function(e) {team_page <- c()})
+
+  player_urls <- team_page %>%
+    rvest::html_nodes("#yw1") %>% rvest::html_nodes(".hauptlink") %>%
+    rvest::html_nodes("a") %>% rvest::html_attr("href") %>%
+    unique() %>%
+    paste0(main_url, .)
+
+  return(player_urls)
+}